requirements.txt
fastapi
uvicorn
motor
jinja2
python-multipart
python-dotenv
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]
starlette
odmantic
email-validator
itsdangerous
pytest
pytest-asyncio
httpx==0.27.0